+++ asphodel/commands/avatar.js
@@ -2,26 +2,29 @@ const Discord = require('discord.js');
 
 module.exports = {
     name: 'avatar',
-    description: 'avatar!',
+    description: "Grab someone's profile picture at the highest resolution.",
     execute(client, message, args) {
         if (args[0] == null) {
             message.reply(`**Incomplete parameters!** This command works in this format:\n\n${prefix}avatar \`[Discord User ID]\` \`[Image Size]\`*(optional)*`);
             return;
         }
-        const theUsersID = args[0];
+        const mention = message.mentions.users.first();
+        const user = mention ? mention.id : args[0];
         const size = args[1] ? parseInt(args[1]) : 4096;
 
-        client.users.fetch(theUsersID).then(myUser => {
+
+        client.users.fetch(user).then(myUser => {
             console.log(myUser.avatarURL());
 
             const response = new Discord.MessageEmbed()
+                .setAuthor(myUser);
 
             if (myUser.displayAvatarURL({ dynamic: true }).endsWith('.gif')) {
                 response.setImage(myUser.avatarURL({ size: size, dynamic: true }))
-                    .setDescription(`Image Size: ${size}\nAnimated: true\nUser ID: ${theUsersID}\nAvatar ID: ${myUser.avatar}`);
+                    .setDescription(`\`\`\`Image Size: ${size}\nAnimated: true\`\`\``);
             } else {
                 response.setImage(myUser.avatarURL({ size: size }))
-                    .setDescription(`Image Size: ${size}\nAnimated: false\nUser ID: ${theUsersID}\nAvatar ID: ${myUser.avatar}`);
+                    .setDescription(`\`\`\`Image Size: ${size}\nAnimated: false\`\`\``);
             }
 
             response.setTimestamp();