+++ asphodel/commands/ping.js
@@ -1,0 +1,15 @@
+const Discord = require('discord.js');
+
+module.exports = {
+    name: 'ping',
+    description: 'Table tennis but with numbers.',
+    async execute(client, message, args) {
+        const exampleEmbed = await new Discord.MessageEmbed()
+            .setColor('#ff0049')
+            .setThumbnail()
+            .setTitle(`🏓 Pong!`)
+            .setDescription(`WebSocket ping is ${client.ws.ping}MS\nMessage sent ping is ${Date.now() - message.createdAt}MS!`)
+            .setTimestamp();
+        const msg = await message.channel.send(exampleEmbed);
+    },
+};