+++ CHANGELOG.md
@@ -1,8 +1,29 @@
 # <strong>ASPHODEL CHANGELOG</strong>
 
-## v0.0.1 prerelease 1 
-> <strong><i>`3f7d8cf0b4f959560a6f18e3362bc699580af303`</i></strong><br/>
-> 03/27/2021 
+## `v0.0.2` - <strong><i>Genesis</i></strong>
+> 29th of March 2021
+
+- Added two new commands!
+    - `ping` tells you the time it takes for the bot to respond. 
+    - `reload` <sup><a href="#fn3" id="ref3">[1]</a></sup>
+- Modified existing commands.
+    - `avatar` can now take in either user IDs and mentioned users.
+    - `help` now tells you the commands available. 
+    - `prune` now has an optional filter to only remove message from a specified user. <a href="https://github.com/salty-sweet/asphodel-bot/issues/1">`#1`</a>
+- Removed `embed.js` entirely.
+
+<strong>Minor Changes</strong>
+- Prefix changed from `/` to `>>`.
+- Modified Changelog so it's less cluttery.
+
+> <sup id="fn3">1. [Command template was added, but currently does nothing.] <a href="#ref3" title="Jump back to footnote 1 in the text.">↩</a></sup><br/>
+
+<br/>
+<br/>
+
+
+## `v0.0.1` prerelease 1
+> 27th of March 2021 
 
 - Added main bot source code.
 - Added assets folder.
@@ -15,13 +36,12 @@
 - Removed `package-lock.json` <sup><a href="#fn2" id="ref2">[2]</a></sup>
 
 > <sup id="fn1">1. [Only for testing embeds. It will probably be gone in the next update.] <a href="#ref1" title="Jump back to footnote 1 in the text.">↩</a></sup><br/>
-> <sup id="fn2">2. [https://i.imgur.com/9vSW9vd.png] <a href="#ref2" title="Jump back to footnote 1 in the text.">↩</a></sup><br/>
+> <sup id="fn2">2. [https://i.imgur.com/9vSW9vd.png] <a href="#ref2" title="Jump back to footnote 2 in the text.">↩</a></sup><br/>
 
 <br/>
 
-## v0.0.1 prerelease 0 
-> <strong><i>`6660c9501a7c99bba6d4ad6dd587cb5a5960a57a`</i></strong><br/>
-> 03/25/2021 
+## `v0.0.1` prerelease 0 
+> 25th of March 2021 
 
 - Initialized NPM on project folder.
 - Added project `README.md` and related assets.